requirements.txt
Django==1.11.17
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
django-suit==0.2.26
djangorestframework==3.9.0
Pillow==5.3.0
psycopg2-binary==2.7.6.1
pytz==2018.7
six==1.12.0
